backend/requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.10.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Vector Database
chromadb>=0.4.0

# LLM & Embeddings (Gemini)
google-generativeai>=0.3.0

# Document Processing & RAG
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
langchain-core>=0.1.0
unstructured>=0.10.0

# File Processing
python-multipart>=0.0.6
pypdf2>=3.0.0
python-docx>=0.8.0

# Text Processing
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
markdown>=3.4.0

# Utilities
requests>=2.31.0
aiofiles>=23.0.0
